requirements.txt
pip==25.2
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-oauth-toolkit==3.0.1
python-decouple==3.8
psycopg2-binary==2.9.10
celery==5.5.3
redis==6.4.0
requests==2.32.5
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-cov==7.0.0
factory-boy==3.3.3
africastalking==2.0
python-jose[cryptography]==3.5.0
authlib==1.6.4
gunicorn==23.0.0
whitenoise==6.11.0
daphne==4.2.1